20200411
1. 오브젝트 매니저 객체 컬링 관리 기능 추가

2. 플레이어 기본총 공격모션 추가

3. 플레이어 부드러운 조작 추가

4. 총알, 총알 충돌 이펙트   추가

5. 총알 지형 충돌 추가

+++ Project_Beom/Project_Beom/AbstractFactory.h
@@ -19,7 +19,7 @@ public:
 	{
 		GameObject* pObj = new SubClass;
 		pObj->Initialize();
-		pObj->SetPosition(posX, posY);
+		pObj->SetPosition((float)posX, (float)posY);
 
 		return pObj;
 	}
